Provide customization
Instead of asking users to modify the hard-coded paths inside the source,
a better way is to provide ability to use Emacs customization facility.

I add the my-whisper customization group with several defcustom variables
that provide choices and checking.  Use M-Tab for directory completion.

The defaults have the same values as the original.
I also modified the formatting of the commands to use the user's choice.