Fix code: no need to pass a quoted lambda
The process sentinel lambda is a closure in Emacs Lisp;
so it has access to the `my-whisper-transcribe` let-bound variables.
Using it this way allows the byte-compiler to check the code.
And it reports that the sentinel `proc` parameter is not used.
Renaming it with a leading underscore silences the warning
and documents the fact that the sentinel does not use the proc value.